refactor to central get_filename
