+++ pliers/stimuli/image.py
@@ -4,12 +4,9 @@ from .base import Stim
 from scipy.misc import imread
 from PIL import Image
 from six.moves.urllib.request import urlopen
-from contextlib import contextmanager
 from scipy.misc import imsave
 import six
 import io
-import os
-import tempfile
 import numpy as np
 
 
@@ -40,13 +37,3 @@ class ImageStim(Stim):
 
     def save(self, path):
         imsave(path, self.data)
-
-    @contextmanager
-    def get_filename(self):
-        if self.filename is None or not os.path.exists(self.filename):
-            tf = tempfile.mktemp() + '.png'
-            imsave(tf, self.data)
-            yield tf
-            os.remove(tf)
-        else:
-            yield self.filename
